Working on exporting JSdoc annotations for IntelliJ IDE documentation
`build.js` will now also generate empty Javascript objects in the
`./lib` directory with correct JSDoc annotations but empty funtions and
undefined variables. The purpose of this is to provide documentation
within IDEs, since many IDEs such as WebStorm and PhpStorm are capable
of parsing the JSDoc annotations and providing documentation within the
IDE. Very useful.